+++ lib/csv_madness.rb
@@ -1,17 +1,8 @@
 require 'csv'
-require 'fun_with_files'
+require 'fun_with_gems'
+# require 'fun_with_version_strings'
 require 'time'        # to use Time.parse to parse cells to get the date
-require 'debugger'
 
-require_relative 'csv_madness/data_accessor_module'
-require_relative 'csv_madness/sheet'
-require_relative 'csv_madness/record'
+lib_dir = __FILE__.fwf_filepath.dirname
+FunWith::Gems.make_gem_fun( "CsvMadness", :require => lib_dir.join( "csv_madness" ) )
 
-
-FunWith::Files::RootPath.rootify( CsvMadness, __FILE__.fwf_filepath.dirname.up )
-
-CsvMadness.class_eval do
-  def self.load( csv, opts = {} )
-    CsvMadness::Sheet.from( csv, opts )
-  end
-end