+++ CHANGELOG.markdown
@@ -1,17 +1,35 @@
+Changelog
+=========
+
+0.0.6
+-----
+
+* CsvMadness::Builder added (easily generate spreadsheets from an array of objects)
+
+
+0.0.5
+-----
+
+* Splitting spreadsheets, creating blank copies of spreadsheets, adding/removing records.
+* Convert records to hashes/arrays.
+* Created a dependency on `fun_with_testing`.
+* CsvMadness.version
+
+
 0.0.4
-=====
+-----
 
 * Fixed serious bug for default column names.  record.middle_name instead of record.middlename
 * Spreadsheet can be re-read from file by calling @sheet.reload_spreadsheet(opts)
 
 
 0.0.3
-=====
+-----
 
 * Feature:  can add and drop columns from spreadsheets
 
 
 0.0.2
-=====
+-----
 
 * Aw, hell.  I don't remember.